Refactored and exposed network indicator and namewithmic status component
